feat: creating the root of the sprout 🌱 (#2)
In this pull request, I have created the core of the Sprout handler to
enable all developers to fully control their functions by configuring
the error handling strategy. This includes the option to utilize logging
based on the new slog package introduced in Go 1.21.

The `FuncMaps()` function continues to exist for backward compatibility
with Sprig.

Additional information will be added to the documentation.