perf: reduce all memory fingerprint (#56)
## Description
Aim to minimize memory allocations as much as possible to alleviate the
burden on the garbage collector in large-scale applications. By
optimizing the way memory is used within the framework, we ensure that
Sprout is not only efficient in its functionality but also in its
resource consumption. This approach contributes to overall better
performance and scalability of applications using Sprout.

## Changes
- Reduce allocations over all functions
- Increase garbage-collector capacity over all functions
- Introduce a robust compare tool to ensure no breaking change are
present between sprig and sprout

## Checklist
- [x] I have read the **CONTRIBUTING.md** document.
- [x] My code follows the code style of this project.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [x] I have updated the documentation accordingly.
- [x] This change requires a change to the documentation on the website.

## Additional Information

Somes functions output are incorrect by definition in Sprig and Sprout
fix that 🎉
A list are presented in `SPRIG_TO_SPROUT_CHANGES_NOTES.md` and needs to
be migrated in official documentation